refactor: overall folder structure
We designed a new structure for our folders.
This commit moves the folders to the new locations.
