+++ lsp/src/service/filetypes/buildspec/activation.ts
