+++ lsp/src/service/filetypes/buildspec/service.ts
