+++ lsp/src/service/registry.ts
