+++ lsp/src/utils/configurationDirectory.ts
