+++ lsp/src/utils/uri/http.ts
